move modules in modules subfolder !! UNTESTED
the goal is, that using 'facebook' in INSTALLED_APPS, no model is loaded to the db
